+++ framework/async/MsgChannel.cpp
@@ -96,7 +96,7 @@ bool idMsgQueue::Get( byte *data, int &size ) {
 	}
 	int sequence;
 	size = ReadShort();
-	sequence = ReadLong();
+	sequence = ReadInt();
 	ReadData( data, size );
 	assert( sequence == first );
 	first++;
@@ -197,10 +197,10 @@ void idMsgQueue::WriteInt( int l ) {
 
 /*
 ===============
-idMsgQueue::ReadLong
+idMsgQueue::ReadInt
 ===============
 */
-int idMsgQueue::ReadLong( void ) {
+int idMsgQueue::ReadInt( void ) {
 	return ReadByte() | ( ReadByte() << 8 ) | ( ReadByte() << 16 ) | ( ReadByte() << 24 );
 }
 
@@ -370,7 +370,7 @@ bool idMsgChannel::ReadMessageData( idBitMsg &out, const idBitMsg &msg ) {
 	out.BeginReading();
 
 	// read acknowledgement of sent reliable messages
-	reliableAcknowledge = out.ReadLong();
+	reliableAcknowledge = out.ReadInt();
 
 	// remove acknowledged reliable messages
 	while( reliableSend.GetFirst() <= reliableAcknowledge ) {
@@ -386,7 +386,7 @@ bool idMsgChannel::ReadMessageData( idBitMsg &out, const idBitMsg &msg ) {
 			common->Printf( "%s: bad reliable message\n", Sys_NetAdrToString( remoteAddress ) );
 			return false;
 		}
-		reliableSequence = out.ReadLong();
+		reliableSequence = out.ReadInt();
 		if ( reliableSequence == reliableReceive.GetLast() + 1 ) {
 			reliableReceive.Add( out.GetData() + out.GetReadCount(), reliableMessageSize );
 		}
@@ -546,7 +546,7 @@ bool idMsgChannel::Process( const netadr_t from, int time, idBitMsg &msg, int &s
 	UpdateIncomingRate( time, msg.GetSize() );
 
 	// get sequence numbers
-	sequence = msg.ReadLong();
+	sequence = msg.ReadInt();
 
 	// check for fragment information
 	if ( sequence & FRAGMENT_BIT ) {
